fix sync error

+++ train.py
@@ -460,25 +460,22 @@ if __name__ == "__main__":
         save_dict["model"] = model.state_dict()
         save_dict["optimizer"] = optimizer.state_dict()
 
-        if epoch == 0 or (epoch+1) % args.test_freq == 0:
-            test_acc = test(args, model, test_loader)
-            # wandb.log({"test_acc":test_acc})
-            # save to best.pt
-            torch.save(save_dict, args.save_root + "backup/last.pth")
-            if test_acc > best_acc:
-                best_acc = test_acc
-                wandb.run.summary["best_accuracy"] = best_acc # upload to wandb
-                wandb.run.summary["best_epoch"] = epoch+1 # upload to wandb
-                if os.path.isfile(args.save_root + "backup/best.pth"):
-                    try:
+        if args.local_rank == 0:
+            if epoch == 0 or (epoch+1) % args.test_freq == 0:
+                test_acc = test(args, model, test_loader)
+                # wandb.log({"test_acc":test_acc})
+                # save to best.pt
+                torch.save(save_dict, args.save_root + "backup/last.pth")
+                if test_acc > best_acc:
+                    best_acc = test_acc
+                    wandb.run.summary["best_accuracy"] = best_acc # upload to wandb
+                    wandb.run.summary["best_epoch"] = epoch+1 # upload to wandb
+                    if os.path.isfile(args.save_root + "backup/best.pth"):
                         os.remove(args.save_root + "backup/best.pth")
-                    except print(0):
-                        pass
-                torch.save(save_dict, args.save_root + "backup/best.pth")
+                    torch.save(save_dict, args.save_root + "backup/best.pth")
 
-        # save to last.pt
-        if os.path.isfile(args.save_root + "backup/last.pth"):
-            try:
+            # save to last.pt
+            if os.path.isfile(args.save_root + "backup/last.pth"):
                 os.remove(args.save_root + "backup/last.pth")
-            except print(0):
-                pass
+        
+        torch.distributed.barrier()
